docker-gif/requirements.txt
Flask
werkzeug
mysql-connector-python
